requirements.txt
telegramify-markdown==0.4.2
requests==2.32.3
boto3==1.37.15
google-generativeai==0.8.4
python-dotenv==1.0.1